+++ packages/configuration/src/index.ts
@@ -8,9 +8,9 @@ const debug = createDebug('@feathersjs/configuration')
 export default function init(schema?: Schema<any> | Validator) {
   const validator: Validator = typeof schema === 'function' ? schema : schema?.validate.bind(schema)
 
-  return (app?: Application) => {
+  return (app?: Application): { [key: string]: unknown } => {
     if (!app) {
-      return config
+      return config as unknown as { [key: string]: unknown }
     }
 
     const configuration: { [key: string]: unknown } = { ...config }
@@ -34,6 +34,6 @@ export default function init(schema?: Schema<any> | Validator) {
       })
     }
 
-    return config
+    return config as unknown as { [key: string]: unknown }
   }
 }