+++ packages/express/test/authentication.test.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
-import { omit } from 'lodash-es'
+import _ from 'lodash'
+const { omit } = _
 import { strict as assert } from 'assert'
 import { default as _axios } from 'axios'
 import { feathers } from '@feathersjs/feathers'