+++ packages/memory/CHANGELOG.md
@@ -47,6 +47,23 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - All packages are now ES modules only
 - **adapter-commons:** Changes the common adapter base class to use `sanitizeQuery` and `sanitizeData`
+## [5.0.43](https://github.com/feathersjs/feathers/compare/v5.0.42...v5.0.43) (2026-03-21)
+
+**Note:** Version bump only for package @feathersjs/memory
+
+## [5.0.42](https://github.com/feathersjs/feathers/compare/v5.0.41...v5.0.42) (2026-03-04)
+
+### Bug Fixes
+
+- Update dependencies ([#3666](https://github.com/feathersjs/feathers/issues/3666)) ([477bf45](https://github.com/feathersjs/feathers/commit/477bf45f9c9dbde77a14a07828aa02300de23ae7))
+
+## [5.0.41](https://github.com/feathersjs/feathers/compare/v5.0.40...v5.0.41) (2026-02-19)
+
+**Note:** Version bump only for package @feathersjs/memory
+
+## [5.0.40](https://github.com/feathersjs/feathers/compare/v5.0.39...v5.0.40) (2026-02-03)
+
+**Note:** Version bump only for package @feathersjs/memory
 
 ## [5.0.39](https://github.com/feathersjs/feathers/compare/v5.0.38...v5.0.39) (2026-01-31)